Revert "Migrate Postwoman to Nuxt.js (full Vue and SCSS support)"
