--- assets/css/fonts.scss
@@ -1,2 +1,0 @@
-// Poppins (Google Fonts)
-@import url("https://fonts.googleapis.com/css?family=Poppins:500,700&display=swap");
