+++ icons/icon-192x192.png
